+++ src/main/java/ksr/zad2/soft/functions/MembershipFunction.java
@@ -18,5 +18,5 @@ public abstract class MembershipFunction {
     }
 
 
-    public abstract Double calculate(double x);
+    public abstract int calculate(double x);
 }
