+++ src/main/java/ksr/zad2/soft/functions/TriangleMembershipFunction.java
@@ -3,7 +3,7 @@ package ksr.zad2.soft.functions;
 import java.util.List;
 
 public class TriangleMembershipFunction extends TrapezoidalMembershipFunction {
-    public TriangleMembershipFunction(double minX, double maxYx, double maxX) {
+    public TriangleMembershipFunction(int minX, int maxYx, int maxX) {
         super(minX, maxYx, maxYx, maxX);
     }
 }
